fixed size_t error in client.c

+++ ffwdh-bench/src/ffwdh-bench-client.c
@@ -27,13 +27,13 @@
     int ch;
     struct in_addr ip;
     struct hostent *hp;
-    
-    struct addrinfo *addrres,hints; 
+
+    struct addrinfo *addrres,hints;
     int error = 0;
 
     const char *ipstr;
     const char *tcpport = "10100";
-    
+
     struct timeval tval_before, tval_after, tval_result;
     float sec;
 
@@ -79,14 +79,14 @@
 
 
     int SocketFD = socket(PF_INET6, SOCK_STREAM, IPPROTO_TCP);
-  
+
     if (-1 == SocketFD) {
       perror("cannot create socket");
       exit(EXIT_FAILURE);
     }
 
 
-    memset(&hints, 0, sizeof(hints)); 
+    memset(&hints, 0, sizeof(hints));
     hints.ai_family = PF_INET6;
 
     error = getaddrinfo(ipstr,tcpport,&hints,&addrres);
@@ -95,13 +95,13 @@
 	fprintf(stderr, "error in getaddrinfo: %s\n", gai_strerror(error));
         exit(EXIT_FAILURE);
     }
-    
+
     if (-1 == connect(SocketFD, addrres->ai_addr,addrres->ai_addrlen)) {
       perror("connect failed");
       close(SocketFD);
       exit(EXIT_FAILURE);
     }
-  
+
     /* take time before */
     gettimeofday(&tval_before, NULL);
 
@@ -109,20 +109,20 @@
 	char buffer[read_buffer_size];
 	read(SocketFD,buffer,read_buffer_size);
     }
-    
+
     /* take time after reading and divide */
     gettimeofday(&tval_after, NULL);
 
-    timersub(&tval_after, &tval_before, &tval_result); 
-    
+    timersub(&tval_after, &tval_before, &tval_result);
+
     /* sec + usec */
     sec = tval_result.tv_sec + (tval_result.tv_usec * 0.000001);
     /* bytes * 8 = bits */
     /* bps = (float) (read_size  / tval_result.tv_usec); */
     bps = (read_size  * 8.0 /  sec);
-    printf("Ready reading %i MB\n", read_size / 1000 / 1000);
+    printf("Ready reading %i MB\n", (int)read_size / 1000 / 1000);
     printf("Reading with  %f MBit/s\n", bps / 1000 / 1000 );
- 
+
     (void)shutdown(SocketFD, SHUT_RDWR);
 
     freeaddrinfo(addrres);
